FXML для domain GUI
